refator: 구매 금액 타입 int에서 long으로 변경

+++ src/main/java/lotto/Application.java
@@ -8,7 +8,7 @@ public class Application {
         User user = new User();
 
         Print.printInsertMoney();
-        int money = Input.readInteger();
+        long money = Input.readLong();
 
         user.payMoney(money);
         user.buyLotto();
