+++ src/main/java/lotto/GameMoney.java
@@ -3,20 +3,20 @@ package lotto;
 public class GameMoney {
     private static final int PRICE_PER_LOTTO = 1000;
 
-    private int money;
+    private long money;
 
-    public GameMoney(int insertMoney) {
+    public GameMoney(long insertMoney) {
         validateAvailableOfPurchase(insertMoney);
         this.money = insertMoney;
     }
 
-    static void validateAvailableOfPurchase(int insertMoney) {
+    static void validateAvailableOfPurchase(long insertMoney) {
         if (insertMoney % PRICE_PER_LOTTO != 0) {
             throw new IllegalArgumentException(Error.ERROR_GAME_MONEY_AVAILABLE.getValue());
         }
     }
 
     public int numberOfPurchases() {
-        return money / PRICE_PER_LOTTO;
+        return (int) money / PRICE_PER_LOTTO;
     }
 }
